+++ frontend/src/store/project/configJsonSlice.tsx
@@ -1,25 +1,24 @@
-import {GetState, SetState} from "zustand";
-import {ProjectState} from "@/store/project/useProjectStore";
-import produce from "immer";
+import type { SetState } from 'zustand';
+import type { ProjectState } from '@/store/project/useProjectStore';
+import produce from 'immer';
 import _ from 'lodash';
-import {message} from "antd";
+import { message } from 'antd';
 
-export type IConfigJsonSlice = {}
+export type IConfigJsonSlice = Record<string, never>;
 
 export interface IConfigJsonDispatchSlice {
   setConfigJson: (value: any) => void;
   setUpgradeType: (value: any) => void;
-};
+}
 
-const ConfigJsonSlice = (set: SetState<ProjectState>,get:GetState<ProjectState>) => ({
+const ConfigJsonSlice = (set: SetState<ProjectState>) => ({
   setConfigJson: (value: any) => set(produce(state => {
-    state.project.configJSON = value
+    state.project.configJSON = value;
   })),
   setUpgradeType: (value: any) => set(produce(state => {
     _.set(state.project, 'configJSON.synchronous.upgradeType', value.upgradeType);
-    message.success("设置成功");
+    message.success('设置成功');
   })),
 });
-
 
 export default ConfigJsonSlice;